+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/AwsCodegenDecorator.kt
@@ -25,6 +25,7 @@ val DECORATORS = listOf(
     AwsFluentClientDecorator(),
     CrateLicenseDecorator(),
     SharedConfigDecorator(),
+    ServiceConfigDecorator(),
     AwsPresigningDecorator(),
     AwsReadmeDecorator(),