+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/AwsFluentClientDecorator.kt
@@ -47,7 +47,13 @@ class AwsFluentClientDecorator : RustCodegenDecorator {
     override fun extras(codegenContext: CodegenContext, rustCrate: RustCrate) {
         val types = Types(codegenContext.runtimeConfig)
         val module = RustMetadata(additionalAttributes = listOf(Attribute.Cfg.feature("client")), public = true)
-        rustCrate.withModule(RustModule("client", module)) { writer ->
+        rustCrate.withModule(
+            RustModule(
+                "client",
+                module,
+                documentation = "Client and fluent builders for calling the service."
+            )
+        ) { writer ->
             FluentClientGenerator(
                 codegenContext,
                 includeSmithyGenericClientDocs = false,
@@ -92,6 +98,7 @@ private class AwsFluentClientExtensions(private val types: Types) {
         writer.rustBlock("impl<C> Client<C, aws_hyper::AwsMiddleware, aws_smithy_client::retry::Standard>") {
             rustTemplate(
                 """
+                /// Creates a client with the given service config and connector override.
                 pub fn from_conf_conn(conf: crate::Config, conn: C) -> Self {
                     let retry_config = conf.retry_config.as_ref().cloned().unwrap_or_default();
                     let client = #{aws_hyper}::Client::new(conn).with_retry_config(retry_config.into());
@@ -104,11 +111,13 @@ private class AwsFluentClientExtensions(private val types: Types) {
         writer.rustBlock("impl Client<aws_smithy_client::erase::DynConnector, aws_hyper::AwsMiddleware, aws_smithy_client::retry::Standard>") {
             rustTemplate(
                 """
+                /// Creates a new client from a shared config.
                 ##[cfg(any(feature = "rustls", feature = "native-tls"))]
                 pub fn new(config: &#{aws_types}::config::Config) -> Self {
                     Self::from_conf(config.into())
                 }
 
+                /// Creates a new client from the service [`Config`](crate::Config).
                 ##[cfg(any(feature = "rustls", feature = "native-tls"))]
                 pub fn from_conf(conf: crate::Config) -> Self {
                     let retry_config = conf.retry_config.as_ref().cloned().unwrap_or_default();