+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/generators/UnionGenerator.kt
@@ -31,6 +31,8 @@ class UnionGenerator(
     }
 
     private fun renderUnion() {
+        writer.documentShape(shape, model)
+
         val unionSymbol = symbolProvider.toSymbol(shape)
         val containerMeta = unionSymbol.expectRustMetadata()
         containerMeta.render(writer)
@@ -51,9 +53,12 @@ class UnionGenerator(
                 if (sortedMembers.size == 1) {
                     Attribute.Custom("allow(irrefutable_let_patterns)").render(this)
                 }
+                rust("/// Tries to convert the enum instance into its #D variant.", unionSymbol)
+                rust("/// Returns `Err(&Self) if it can't be converted.` ")
                 rustBlock("pub fn as_$funcNamePart(&self) -> std::result::Result<&#T, &Self>", memberSymbol) {
                     rust("if let ${unionSymbol.name}::$variantName(val) = &self { Ok(&val) } else { Err(&self) }")
                 }
+                rust("/// Returns true if the enum instance is the `${unionSymbol.name}` variant.")
                 rustBlock("pub fn is_$funcNamePart(&self) -> bool") {
                     rust("self.as_$funcNamePart().is_ok()")
                 }