+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/generators/config/IdempotencyTokenProviderCustomization.kt
@@ -27,17 +27,19 @@ class IdempotencyTokenProviderCustomization : NamedSectionGenerator<ServiceConfi
             ServiceConfig.BuilderImpl -> writable {
                 rust(
                     """
-            pub fn make_token(mut self, make_token: impl Into<#T::IdempotencyTokenProvider>) -> Self {
-                self.make_token = Some(make_token.into());
-                self
-            }
-            """,
+                    /// Sets the idempotency token provider to use for service calls that require tokens.
+                    pub fn make_token(mut self, make_token: impl Into<#T::IdempotencyTokenProvider>) -> Self {
+                        self.make_token = Some(make_token.into());
+                        self
+                    }
+                    """,
                     RuntimeType.IdempotencyToken
                 )
             }
             ServiceConfig.BuilderBuild -> writable {
                 rust("make_token: self.make_token.unwrap_or_else(#T::default_provider),", RuntimeType.IdempotencyToken)
             }
+            else -> writable { }
         }
     }
 }
@@ -61,6 +63,7 @@ impl Builder {
         Self::default()
     }
 
+    /// Sets the idempotency token provider to use for service calls that require tokens.
     pub fn make_token(
         mut self,
         make_token: impl crate::idempotency_token::MakeIdempotencyToken + 'static,