+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/generators/error/ErrorGenerator.kt
@@ -78,13 +78,15 @@ class ErrorGenerator(
         writer.rustBlock("impl ${symbol.name}") {
             val retryKindWriteable = shape.modeledRetryKind(error)?.writable(symbolProvider.config().runtimeConfig)
             if (retryKindWriteable != null) {
+                rust("/// Returns `Some(${errorKindT.name})` if the error is retryable. Otherwise, returns `None`.")
                 rustBlock("pub fn retryable_error_kind(&self) -> #T", errorKindT) {
                     retryKindWriteable(this)
                 }
             }
             rust(
                 """
-            pub fn message(&self) -> Option<&str> { $message }
+                /// Returns the error message.
+                pub fn message(&self) -> Option<&str> { $message }
                 """
             )
         }